+++ src/utils/db/updateProjectUpdatedAt.ts
@@ -1,0 +1,10 @@
+import { prisma } from '../prisma';
+
+export const updateProjectUpdatedAt = async (id?: string | null) => {
+    if (!id) return;
+
+    return prisma.project.update({
+        where: { id },
+        data: { id },
+    });
+};
