+++ src/utils/worker/index.ts
@@ -2,13 +2,12 @@ import { PrismaClient } from '@prisma/client';
 import * as Sentry from '@sentry/nextjs';
 import parser from 'cron-parser';
 
-import { jobKind, jobState } from './create';
+import { jobKind, jobState, defaultJobDelay } from './create';
 import * as resolve from './resolve';
 
 const prisma = new PrismaClient();
 const queueInterval = process.env.WORKER_JOBS_INTERVAL ? parseInt(process.env.WORKER_JOBS_INTERVAL, 10) : 3000;
 const retryLimit = process.env.WORKER_JOBS_RETRY ? parseInt(process.env.WORKER_JOBS_RETRY, 10) : 3;
-const defaultJobDelay = process.env.WORKER_JOBS_DELAY ? parseInt(process.env.WORKER_JOBS_DELAY, 10) : 1000;
 
 // eslint-disable-next-line no-console
 console.log('Worker started successfully');
@@ -53,6 +52,10 @@ console.log('Worker started successfully');
                         }
                     }
 
+                    if (job.delay && Date.now() - new Date(job.createdAt).valueOf() < job.delay) {
+                        return;
+                    }
+
                     setTimeout(async () => {
                         await prisma.job.update({ where: { id: job.id }, data: { state: jobState.pending } });
                     }, 0);
@@ -71,9 +74,14 @@ console.log('Worker started successfully');
                                 setTimeout(async () => {
                                     await prisma.job.update({
                                         where: { id: job.id },
-                                        data: { state: jobState.scheduled, error: error?.message, retry },
+                                        data: {
+                                            state: jobState.scheduled,
+                                            error: error?.message,
+                                            retry,
+                                            delay: defaultJobDelay * retry,
+                                        },
                                     });
-                                }, retry * defaultJobDelay);
+                                }, 0);
                             } else {
                                 Sentry.captureException(error, {
                                     fingerprint: ['worker', 'resolve', 'retry'],
@@ -85,7 +93,7 @@ console.log('Worker started successfully');
                                 await prisma.job.delete({ where: { id: job.id } });
                             }
                         }
-                    }, job.delay || defaultJobDelay);
+                    }, 0);
                 }
             });
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
